requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
sentence-transformers
huggingface-hub
python-dotenv
